requirements.txt
langchain
openai
python-dotenv
fastapi
bs4
unstructured